agritech-ai-analyzer/requirements.txt
streamlit==1.29.0
torch==2.2.0
torchvision==0.17.0
tensorflow==2.16.1
keras==3.0.1
ultralytics==8.1.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2
scikit-learn==1.3.2
joblib==1.2.0